fix: extract generateAlias to own module (react-refresh lint) (#PAT-104)
LibraryDefinitionPicker.tsx mixing a named non-component export with
its default component export triggers react-refresh/only-export-components
on CI (treated as error via --max-warnings 0).

Move generateAlias to utils/libraryAlias.ts. Picker imports it; test
imports from the new module directly.
